add new TIR kernel
